requirements.txt
Django==1.10.6
ipython==4.2.0
djangorestframework==3.5.3
django-allauth==0.30.0
django-rest-auth==0.9.0
django-rest-swagger==2.1.1
factory-boy==2.8.1
flake8
